Panier fini

+++ js/script.js
@@ -27,34 +27,50 @@ input.onkeyup = (e)=>{
 //A enlever plus tard
 localStorage.setItem('items', ['Item_1', 'Item_2', 'Item_3']);
 
-function showTrolley() {
+let continueToShowDropdown = false;
+
+function showCart() {
     const items = localStorage.getItem("items").split(',');
-    const html = [];
-    for (let i = 0 ; i < items.length ; i++) {
-        html.push(
-            "<div class='trolley_item' id='" + items[i] + "'>" + 
-                "<img src='images/logo.jpg'>" +
-                "<p>" + items[i] + "<br>10,99€</p>" +
-                "<p>Qte : 2</p>" +
-                "<i class='fa-solid fa-xmark' onclick='deleteItem(\"" + items[i] + "\")'></i>" +
-            "</div>"
-        );
+    if (items.length === 0 || (items.length === 1 && items[0] == '')) {
+        document.querySelector("#cart_items").innerHTML = "<p style='width: max-content'>Aucun item dans le panier</p>";
+        document.querySelector("#cart_items").style.display = document.querySelector("#cart_items").style.display == "block" ? "none" : "block";
+    } else {
+        const html = [];
+        for (let i = 0 ; i < items.length ; i++) {
+            html.push(
+                "<li class='cart_item' id='" + items[i] + "'>" + 
+                    "<img src='images/logo.jpg'>" +
+                    "<p>" + items[i] + "<br>10,99€</p>" +
+                    "<p style='width: max-content'>Qte : 2</p>" +
+                    "<i class='fa-solid fa-xmark' onclick='deleteItem(\"" + items[i] + "\")'></i>" +
+                "</li>"
+            );
+        }
+        document.querySelector("#cart_items").innerHTML = html.join("");
+        document.querySelector("#cart_items").style.display = document.querySelector("#cart_items").style.display == "block" ? "none" : "block";
     }
-    document.querySelector("#trolley_items").innerHTML = html.join("");
-    document.querySelector("#trolley_items").style.display = document.querySelector("#trolley_items").style.display == "block" ? "none" : "block";
+    
 }
 
 function deleteItem(item) {
     const items = localStorage.getItem("items").split(',');
+    console.log(items);
     const index = items.indexOf(item);
     if (index == -1) return;
     items.splice(index, 1);
     localStorage.setItem("items", items);
-    document.querySelector("#trolley_items #" + item).remove();
+    document.querySelector("#cart_items #" + item).remove();
+    continueToShowDropdown = true;
+    document.querySelector("#cart_count").innerHTML = items.length;
+    if (items.length == 0) {
+        document.querySelector("#cart_items").innerHTML = "<p style='width: max-content'>Aucun item dans le panier</p>";
+        document.querySelector("#cart_count").style.display = "none";
+    }
 }
 
 document.addEventListener("click", function (event) {
-    if (!document.querySelector("#icons").contains(event.target)) {
-        document.querySelector("#trolley_items").style.display = "none";
+    if (!document.querySelector("#icons").contains(event.target) && !continueToShowDropdown) {
+        document.querySelector("#cart_items").style.display = "none";
     }
+    continueToShowDropdown = false;
 });
